Front-End stage one
This is stage one of the front end, including the home page, predict page, and about page. The programming languages/tools used in stage one are HTML, CSS, Javascript, and Jquery. Stage one does not have a user interaction with the data and models as they will be implemented in front-end stage two.